refactor: clean up file diff iterator logic (#120)
* fix: make all entrie scores start at 1, fix incorrect indecisions

* test: add more hoard_item tests

* chore: update coverage flags for Rust 1.60.0

* refactor!: replace diff iterator logic

This also removes the "Recreated" type, since it is too difficult to
tell the difference between it and "Created".

* fix: fix clippy errors/warnings

* style: run cargo fmt

* chore: ignore lines containing unreachable!()

* test: rewrite hoard diff tests for easier expansion

* fix: missing/incorrect diff conditions

* fix: remove debug println

* fix: derive Ord when deriving PartialOrd

* refactor: make FileContent available to rest of crate

* test: more hoard diff tests with new macros

* refactor: introduce CachedHoardItem

* fix: fix a few failing tests

* fix: remove debug println

* test: add more hoard diff tests

* fix!: correctly detect files deleted in both places

- Added checks in most recent logs for files that existed before but not
  now.
- Added `HoardFileDiff::Nonexistent` to indicate a file that is listed
  in the config but has never existed in Hoard.
- Added `ItemOperation::DoesNotExist` to correspond with the above.
- Because of these changes, a couple of "unreachable" branches became
  reachable, so that got fixed too.

* refactor: merge expected and unexpected diffs into one method

* refactor: remove references to permissions

Permissions checks were among the last few things to get test coverage,
but there is not a reliable way to maintain permissions across
filesystems and/or operating systems, particularly when Windows is
involved.

This removes permissions support. In the future, there will be a change
to apply permissions on restore.

* refactor: remove dead code

* style: run cargo fmt

* chore: clean up test code, remove unused dependency

+++ src/command/backup_restore.rs
@@ -1,5 +1,6 @@
+use crate::checkers::history::operation::ItemOperation;
 use crate::checkers::{history::operation::OperationImpl, Checkers, Error as ConsistencyError};
-use crate::hoard::iter::{Error as IterError, ItemOperation};
+use crate::hoard::iter::Error as IterError;
 use crate::hoard::{Direction, Hoard};
 use crate::newtypes::HoardName;
 use crate::paths::{HoardPath, RelativePath};
@@ -105,14 +106,19 @@ fn backup_or_restore<'a>(
                         Direction::Backup => file.hoard_path().as_ref(),
                         Direction::Restore => file.system_path().as_ref(),
                     };
-                    tracing::debug!("deleting {}", to_remove.display());
-                    if let Err(err) = fs::remove_file(to_remove) {
-                        tracing::error!("failed to delete {}: {}", to_remove.display(), err);
-                        return Err(Error::IO(err));
+                    if to_remove.exists() {
+                        tracing::debug!("deleting {}", to_remove.display());
+                        if let Err(err) = fs::remove_file(to_remove) {
+                            tracing::error!("failed to delete {}: {}", to_remove.display(), err);
+                            return Err(Error::IO(err));
+                        }
                     }
                 }
                 ItemOperation::Nothing(file) => {
                     tracing::debug!("file {} is unchanged", file.system_path().display());
+                }
+                ItemOperation::DoesNotExist(file) => {
+                    tracing::trace!("file {} does not exist", file.system_path().display());
                 }
             }
         }
